Update type line parser
Add proper structs and parsers for more common types. I extend the
prompt options to a vec of options to make this better used as a
formatter. I don't want to break on invalid input if its close enough,
I'd rather tag the objects with warning and errors for formatting issues
and then a fmt'er for spitting printing the struct out like you would
find in a Kconfig file