We can rebuilt it...
We can make it stronger. Also delete a bunch of code.

Key points this commit are proper hex and signed int parsing as well as
more structs.

I am 95% sure what I want to do is far more trait-like and far lest
struct-y. But also ive got to interop with kernelfied C and Makefiles.
So its really a toss of if I want to go 1:1 or make a new structure all
together.