Adding "ConstantSymbol"
According to the KConfig language docs, what happens between quotes is
called a Constant Symbol (in contrast to a config option which is a
Non-Constant Symbol). Its mostly correct, but I may enable the
parentheses and dollarparentheses varaints and parse macros with this
structure as well.

I will need to rename the struct at that point to avoid confusion

The Symbol struct now explictily declares all chars it accepts. I
thought about chaining parsers for it, but for now I will keep it is_a